src: add kNoBrowserGlobals flag for Environment
