+++ lib/internal/util.js
@@ -466,6 +466,26 @@ function createDeferredPromise() {
   return { promise, resolve, reject };
 }
 
+// https://heycam.github.io/webidl/#define-the-operations
+function defineOperation(target, name, method) {
+  ObjectDefineProperty(target, name, {
+    writable: true,
+    enumerable: true,
+    configurable: true,
+    value: method
+  });
+}
+
+// https://heycam.github.io/webidl/#es-interfaces
+function exposeInterface(target, name, interfaceObject) {
+  ObjectDefineProperty(target, name, {
+    writable: true,
+    enumerable: false,
+    configurable: true,
+    value: interfaceObject
+  });
+}
+
 let _DOMException;
 const lazyDOMExceptionClass = () => {
   _DOMException ??= internalBinding('messaging').DOMException;
@@ -484,8 +504,10 @@ module.exports = {
   createClassWrapper,
   createDeferredPromise,
   decorateErrorStack,
+  defineOperation,
   deprecate,
   emitExperimentalWarning,
+  exposeInterface,
   filterDuplicateStrings,
   getConstructorOf,
   getSystemErrorMap,
